refactoring and moved tests
