+++ Cargo.toml
@@ -155,6 +155,7 @@ reqwest = { version = "0.12", default-features = false }
 anyhow = "1.0"
 criterion = { version = "0.5.1", features = ["html_reports", "tokio"] }
 rayon = "1.10"
+rstest = "0.22.0"
 tokio = { version = "1.37", features = [
     "rt",
     "rt-multi-thread",
